PID
Pid works pretty good at power 80.

+++ Arduino/ControlDeveloper_PID_Tests/QRE1113_Driver.h
@@ -8,7 +8,7 @@ class QRE1113
 public:
 	int sensorPin = 0;
 	int currVal = 0, lastVal = 0;
-	int currSteps = 0, lastSteps = 0;
+	long currSteps = 0, lastSteps = 0;
 
 	//This mean that sensors have to be calibrated on every car
 	int minUnder = 2500; //under which value will be considered 0 (0-4095)
